Update local MySQL DB configuration
Using ./configure_repositories setup-local-mysql ../db-migrations.

The main change here is that the latest image of serlo-mysql-database is always pulled, so that the latest version of the image is always used.
In this case, the goal was to get the latest changes related to the fixed encoding errors, where the default-character-set is set to `utf8mb4`.